refactoring according to code reviewer's comments
